Switch from auto_now_add=True to default=now
I've run in the following problem a few times in tests:

* I create a model instance
* set its "created" field to a specific value
* I save the model instance
* I write testcase logic which relies on that specific "created" value

The testcase fails, because, with auto_now_add=True, Django
overwrites the created field. I can work around this by:

* Create and save a model instance
* Save it
* Set the created field to my desired value
* Save it again

But this is annoying to do, and annoying to troubleshoot
– it's easy to forget about the auto_now_add behaviour.

So I'm replacing auto_now_add=True with
default=django.utils.timezone.now.